requirements.txt
gunicorn
Flask
werkzeug
sslcrypto
pyTelegramBotAPI
datetime
tg-logger
pytz
langchain
PyPDF2
pinecone-client
pypdf
openai
cohere
chromadb